Use AsyncSequence of NSCalendarDayChanged Notificaitons
- Fixes #293
